# readme/api-docs/api-docs-openapi-3.1.0.json
{"openapi":"3.1.0","info":{"title":"Platform API","description":"Platform API","version":"1.0"},"servers":[{"url":"http://localhost:8080/api","description":"Generated server url"}],"paths":{"/user/login/refreshToken":{"post":{"tags":["user-login-controller"],"summary":"user refreshToken to get new tokens","description":"don't change the current path name, as there is special logic with it in JwtAuthenticationFilter.java\n","operationId":"refreshToken","parameters":[{"name":"refreshToken","in":"header","required":true,"schema":{"type":"string","default":"Bearer "}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/user/login/outLogin":{"post":{"tags":["user-login-controller"],"summary":"logout","description":"logout and redirect to login page","operationId":"outLogin","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"example":"Set this test token in OpenApiConfig.java"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/user/login/login":{"post":{"tags":["user-login-controller"],"summary":"login","description":"login and return jwt accessToken and refreshToken","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/demo/rule/testPost":{"post":{"tags":["rule-controller"],"summary":"for your test, you can return any JSON to front as you want","operationId":"testPost","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"example":"Set this test token in OpenApiConfig.java"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/demo/rule/insert":{"post":{"tags":["rule-controller"],"summary":"insert the rule record","operationId":"insert","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"example":"Set this test token in OpenApiConfig.java"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RuleInsertDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/demo/message/errorMessageDemo":{"post":{"tags":["message-controller"],"summary":"test error message","operationId":"errorMessageDemo","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorMessageDemoDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/demo/file/uploadFileTest":{"post":{"tags":["file-controller"],"summary":"test file upload","operationId":"uploadFileTest","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"example":"Set this test token in OpenApiConfig.java"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/user/login/currentUser":{"get":{"tags":["user-login-controller"],"summary":"get current user info","description":"response JSON below is from Ant Design Pro\n\n# success response\n{\"success\":true,\"data\":{\"name\":\"Serati Ma\",\"avatar\":\"https://gw.alipayobjects.com/zos/antfincdn/XAosXuNZyF/BiazfanxmamNRoxxVxka.png\",\"userid\":\"00000001\",\"email\":\"antdesign@alipay.com\",\"signature\":\"海纳百川，有容乃大\",\"title\":\"交互专家\",\"group\":\"蚂蚁金服－某某某事业群－某某平台部－某某技术部－UED\",\"tags\":[{\"key\":\"0\",\"label\":\"很有想法的\"},{\"key\":\"1\",\"label\":\"专注设计\"},{\"key\":\"2\",\"label\":\"辣~\"},{\"key\":\"3\",\"label\":\"大长腿\"},{\"key\":\"4\",\"label\":\"川妹子\"},{\"key\":\"5\",\"label\":\"海纳百川\"}],\"notifyCount\":12,\"unreadCount\":11,\"country\":\"China\",\"access\":\"admin\",\"geographic\":{\"province\":{\"label\":\"浙江省\",\"key\":\"330000\"},\"city\":{\"label\":\"杭州市\",\"key\":\"330100\"}},\"address\":\"西湖区工专路 77 号\",\"phone\":\"0752-268888888\"}}\n\n# fail response (we don't need this fail response, as the front-end will be redirected to login page, rather than showing this message)\n{\"data\":{\"isLogin\":false},\"errorCode\":\"401\",\"errorMessage\":\"请先登录！\",\"success\":true}\n","operationId":"currentUser","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"example":"Set this test token in OpenApiConfig.java"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/demo/rule/selectPage":{"get":{"tags":["rule-controller"],"summary":"select the rule record by page","description":"you can use \"name\" to test different exception, for example, input \"1\" to test WARN_MESSAGE, check the code to see why\n","operationId":"selectPage","parameters":[{"name":"current","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"pageSize","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"name","in":"query","required":true,"schema":{"type":"string"}},{"name":"Authorization","in":"header","required":true,"schema":{"example":"Set this test token in OpenApiConfig.java"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/demo/file/downloadExcelTest":{"get":{"tags":["file-controller"],"summary":"test download Excel","operationId":"downloadExcelTest","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"example":"Set this test token in OpenApiConfig.java"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"LoginDTO":{"type":"object","description":"Table name: user_account<br/>Description: the login account and the basic information of the user\n","properties":{"username":{"type":"string","description":"Username for login, must be unique\n","minLength":1},"password":{"type":"string","description":"Password with encryption\n","minLength":1}},"required":["password","username"]},"RuleInsertDTO":{"type":"object","description":"Table name: demo_rule<br/>Description: Table to store demo records of rule\n","properties":{"name":{"type":"string","description":"Name of the item or entity\n"},"desc":{"type":"string","description":"Description of the rule\n"}},"required":["name"]},"ErrorMessageDemoDTO":{"type":"object","description":"error message demo\n","properties":{"errorShowType":{"type":"integer","format":"int32","description":"Used to control the error message display type at front end </br>\nRefer to ErrorShowTypeEnum: </br>\n        SILENT(0),\n        WARN_MESSAGE(1),\n        ERROR_MESSAGE(2),\n        NOTIFICATION(3),\n","maximum":3,"minimum":0}},"required":["errorShowType"]}}}}